Move Jackson Guice adapters into io.druid * Removes access to protected methods in com.fasterxml * Eliminates druid-common's use of foreign package com.fasterxml
